fix(configure-reporting): refactor and improve configure reporting commands parsers (#28)
* fix(configure-reporting): refactor and improve configure reporting commands parsers

There was a bug that caused one of two configure reporting records being send in one `configureReporting` command to be ignored. The last field of the record is `minChange` which is a variable length buffer. This would cause the parser to ignore the rest of the command. To accommodate for this a custom parser was created which handles converting `configureReporting` commands to and from a buffer. Additionally, since the `readReportingConfiguration` command response has a very similar structure it was refactored to also use the `ZNPConfigureReportingStruct` data type (with a `withStatus` property, since this property is only available in the `readReportingConfiguration` command response).

Tests were added for both `configureReporting` commands and `readReportingConfiguration` responses.

* chore(deps): bump @athombv/data-types

* refactor: configure reporting `withStatus` parameter